TECH: Rename connectors
