+++ package.json
@@ -59,8 +59,8 @@
     "start": "export BROWSER=none && PORT=3300 react-scripts start",
     "start-browser": "PORT=3301 react-scripts start",
     "start-electron": "export ELECTRON_START_URL=http://localhost:3300 && electron .",
-    "build-and-start-electron": "yarn build && yarn build-electron && electron .",
     "start-electron-dev-tools": "DEV_TOOLS=true export ELECTRON_START_URL=http://localhost:3300 && cross-env nodemon --exec electron .",
+    "build-and-start-electron": "yarn build && yarn build-electron && electron .",
     "start-win": "set BROWSER=none && react-scripts start",
     "start-electron-win": "set ELECTRON_START_URL=http://localhost:3300 && electron .",
     "clean": "rm -rf build dist",
