+++ src/react/App.js
@@ -1,11 +1,12 @@
 import React from 'react'
+
 import { reducerWithMiddleware } from './utils/middleware'
 import { StoreProvider } from './store/StoreContext'
 import { reducer } from './store/reducer'
 import { initialState } from './store/initialState'
 
 import ButtonBar from './features/buttons/ButtonBar'
-import Connector from './features/connector/Connector'
+import Connectors from './features/connectors/Connectors'
 import Filters from './features/filters/Filters'
 import Navigation from './features/navigation/Navigation'
 import View from './View'
@@ -15,7 +16,7 @@ const App = () => (
     <Navigation />
     <ButtonBar />
     <View />
-    <Connector />
+    <Connectors />
     <Filters />
   </div>
 )
