+++ src/react/features/connectors/Connectors.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import FolderConnector from './folder/FolderConnector'
+
+const Connectors = () => {
+  return (
+    <>
+      <FolderConnector/>
+    </>
+  )
+}
+
+export default Connectors
