+++ src/react/features/connectors/actionsTypes.js
