+++ src/react/features/connectors/folder/FolderConnector.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { useButtons } from '../buttons/useButtons'
+import { useButtons } from '../../buttons/useButtons'
 
 import { useGetRoom } from './hooks/useGetRoom'
 import { useSaveRoom } from './hooks/useSaveRoom'
@@ -8,7 +8,7 @@ import { useOpenRoom } from './hooks/useOpenRoom'
 
 import { buttons } from './buttons'
 
-const Connector = () => {
+const FolderConnector = () => {
   useButtons(buttons)
 
   useGetRoom()
@@ -18,4 +18,4 @@ const Connector = () => {
   return <div data-testid="connector" />
 }
 
-export default Connector
+export default FolderConnector
