+++ src/react/features/connectors/folder/hooks/mockResponses/GET_ROOM.json
