+++ src/react/features/connectors/reducers.js
