+++ src/react/store/reducer.js
@@ -1,5 +1,5 @@
 import buttons from '../features/buttons/reducers'
-import connector from '../features/connector/reducers'
+import connector from '../features/connectors/reducers'
 import filters from '../features/filters/reducers'
 import lenses from '../features/lenses/reducers'
 import navigation from '../features/navigation/reducers'
