+++ src/react/utils/cucumber/support/data/expectLinksInDiory.js
@@ -1,5 +1,5 @@
 export const expectLinksInDiory = (idPrefix, id, links) => {
-  const { diograph } = require('../../../../features/connector/hooks/mockResponses/GET_ROOM.json')
+  const { diograph } = require('../../../../features/connectors/folder/hooks/mockResponses/GET_ROOM.json')
   const diory = diograph[idPrefix + id]
   expect(Object.keys(diory.links).length).toBe(links)
 }
